v1.0.0 = Initial Version
Converted the Bash version into Python...because...why not?  (Mainly a reason to learn more Python.)

This is a quick "Patch Management Framework" that I threw together quickly one day for a need (cough zoom cough).

Not a big fan of the built in patch "features" in Jamf Pro -- they leave a bit to be desired.  This, more or less, combines the functionality of the two approaches (Standard Policies and Patch Policies) and makes them appear more "user friendly."