v1.1.0 = Resolved issue with patching
AutoCAD patches will not install from command line, only via GUI.  This because they have a script that runs on launch of the .pkg (via GUI) that locates the base .app that needs to be upgraded, and then creates a symlink to that .app in /tmp.

This updated "fixes" this issue with the usual trickery.