v1.3.0 = Use official installation method
This version of the script uses the official installation method.

The previous installation method (that worked for v2021 and v2022) stopped working with v2023.  It was really a hack work around in the first palce.

The only problem with this method is, it does not work when it is executed from a .pkg's postinstall script for some odd reason.  So the script will need to be ran separately from the Install App's distribution method.